@@ -11,22 +11,13 @@ import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
 import io.grpc.netty.GrpcSslContexts;
 import io.grpc.netty.NettyChannelBuilder;
+import io.opentelemetry.exporter.otlp.internal.TlsUtil;
 import io.opentelemetry.sdk.common.CompletableResultCode;
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.security.KeyManagementException;
-import java.security.KeyStore;
-import java.security.KeyStoreException;
-import java.security.NoSuchAlgorithmException;
-import java.security.cert.CertificateException;
-import java.security.cert.CertificateFactory;
-import java.security.cert.X509Certificate;
 import java.util.concurrent.TimeUnit;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import javax.net.ssl.SSLContext;
 import javax.net.ssl.SSLException;
-import javax.net.ssl.TrustManagerFactory;
+import javax.net.ssl.X509TrustManager;
 
 /**
  * Utilities for working with gRPC channels.
@@ -50,7 +41,7 @@ public final class ManagedChannelUtil {
     requireNonNull(managedChannelBuilder, "managedChannelBuilder");
     requireNonNull(trustedCertificatesPem, "trustedCertificatesPem");
 
-    TrustManagerFactory tmf = trustManagerFactory(trustedCertificatesPem);
+    X509TrustManager tmf = TlsUtil.trustManager(trustedCertificatesPem);
 
     // gRPC does not abstract TLS configuration so we need to check the implementation and act
     // accordingly.
@@ -71,43 +62,11 @@ public final class ManagedChannelUtil {
         .equals("io.grpc.okhttp.OkHttpChannelBuilder")) {
       io.grpc.okhttp.OkHttpChannelBuilder okHttpBuilder =
           (io.grpc.okhttp.OkHttpChannelBuilder) managedChannelBuilder;
-      SSLContext sslContext;
-      try {
-        sslContext = SSLContext.getInstance("TLS");
-        sslContext.init(null, tmf.getTrustManagers(), null);
-      } catch (NoSuchAlgorithmException | KeyManagementException e) {
-        throw new SSLException("Could not build SSLContext.", e);
-      }
-      okHttpBuilder.sslSocketFactory(sslContext.getSocketFactory());
+      okHttpBuilder.sslSocketFactory(TlsUtil.sslSocketFactory(tmf));
     } else {
       throw new SSLException(
           "TLS certificate configuration not supported for unrecognized ManagedChannelBuilder "
               + managedChannelBuilder.getClass().getName());
-    }
-  }
-
-  private static TrustManagerFactory trustManagerFactory(byte[] trustedCertificatesPem)
-      throws SSLException {
-    requireNonNull(trustedCertificatesPem, "trustedCertificatesPem");
-    try {
-      KeyStore ks = KeyStore.getInstance(KeyStore.getDefaultType());
-      ks.load(null);
-
-      ByteArrayInputStream is = new ByteArrayInputStream(trustedCertificatesPem);
-      CertificateFactory factory = CertificateFactory.getInstance("X.509");
-      int i = 0;
-      while (is.available() > 0) {
-        X509Certificate cert = (X509Certificate) factory.generateCertificate(is);
-        ks.setCertificateEntry("cert_" + i, cert);
-        i++;
-      }
-
-      TrustManagerFactory tmf =
-          TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
-      tmf.init(ks);
-      return tmf;
-    } catch (CertificateException | KeyStoreException | IOException | NoSuchAlgorithmException e) {
-      throw new SSLException("Could not build TrustManagerFactory from trustedCertificatesPem.", e);
     }
   }
 
