@@ -35,7 +35,8 @@ class OpenTelemetryTextFormatImpl extends TextFormat {
   public <C> void inject(SpanContext spanContext, C carrier, Setter<C> setter) {
     io.opentelemetry.api.trace.SpanContext otelSpanContext = mapSpanContext(spanContext);
     Context otelContext = Context.current().with(Span.wrap(otelSpanContext));
-    propagator.inject(otelContext, carrier, setter::put);
+    // Use explicit lambda instead of method reference for nullness check.
+    propagator.inject(otelContext, carrier, (c, key, value) -> setter.put(c, key, value));
   }
 
   @Override
