@@ -68,8 +68,7 @@ public class OtlpPipelineStressTest {
   @Container
   public static GenericContainer<?> collectorContainer =
       new GenericContainer<>(
-              DockerImageName.parse(
-                  "ghcr.io/open-telemetry/java-test-containers:otel-collector-dev"))
+              DockerImageName.parse("ghcr.io/open-telemetry/opentelemetry-java/otel-collector"))
           .withNetwork(network)
           .withNetworkAliases("otel-collector")
           .withExposedPorts(OTLP_RECEIVER_PORT)
@@ -94,7 +93,7 @@ public class OtlpPipelineStressTest {
   @Container
   public static GenericContainer<?> toxiproxyContainer =
       new GenericContainer<>(
-              DockerImageName.parse("ghcr.io/open-telemetry/java-test-containers:toxiproxy"))
+              DockerImageName.parse("ghcr.io/open-telemetry/opentelemetry-java/toxiproxy"))
           .withNetwork(network)
           .withNetworkAliases("toxiproxy")
           .withExposedPorts(TOXIPROXY_CONTROL_PORT, COLLECTOR_PROXY_PORT)
