@@ -28,7 +28,7 @@ class JaegerRemoteSamplerIntegrationTest {
 
   @Container
   public static GenericContainer<?> jaegerContainer =
-      new GenericContainer<>("ghcr.io/open-telemetry/java-test-containers:jaeger")
+      new GenericContainer<>("ghcr.io/open-telemetry/opentelemetry-java/jaeger")
           .withCommand("--sampling.strategies-file=/sampling.json")
           .withExposedPorts(COLLECTOR_PORT, QUERY_PORT, HEALTH_PORT)
           .waitingFor(Wait.forHttp("/").forPort(HEALTH_PORT))
