@@ -111,7 +111,11 @@ class WorkerExporter {
       }
     }
     exportCurrentBatch(batch);
-    flushSignal.get().succeed();
-    flushSignal.set(null);
+    CompletableResultCode flushResult = flushSignal.get();
+    // In practice shouldn't be null but easiest way to work with the null checker.
+    if (flushResult != null) {
+      flushResult.succeed();
+      flushSignal.set(null);
+    }
   }
 }
