@@ -8,13 +8,14 @@ package io.opentelemetry.sdk.metrics;
 import io.opentelemetry.api.metrics.Meter;
 import io.opentelemetry.api.metrics.MeterBuilder;
 import io.opentelemetry.sdk.internal.ComponentRegistry;
+import javax.annotation.Nullable;
 
 class SdkMeterBuilder implements MeterBuilder {
 
   private final ComponentRegistry<SdkMeter> registry;
   private final String instrumentationName;
-  private String instrumentationVersion;
-  private String schemaUrl;
+  @Nullable private String instrumentationVersion;
+  @Nullable private String schemaUrl;
 
   SdkMeterBuilder(ComponentRegistry<SdkMeter> registry, String instrumentationName) {
     this.registry = registry;
