@@ -7,7 +7,7 @@ package io.opentelemetry.sdk.metrics.internal.state;
 
 import com.google.auto.value.AutoValue;
 import io.opentelemetry.sdk.common.Clock;
-import io.opentelemetry.sdk.metrics.exemplar.ExemplarSampler;
+import io.opentelemetry.sdk.metrics.exemplar.ExemplarFilter;
 import io.opentelemetry.sdk.metrics.internal.view.ViewRegistry;
 import io.opentelemetry.sdk.resources.Resource;
 import javax.annotation.concurrent.Immutable;
@@ -22,9 +22,9 @@ import javax.annotation.concurrent.Immutable;
 @Immutable
 public abstract class MeterProviderSharedState {
   public static MeterProviderSharedState create(
-      Clock clock, Resource resource, ViewRegistry viewRegistry, ExemplarSampler sampler) {
+      Clock clock, Resource resource, ViewRegistry viewRegistry, ExemplarFilter exemplarFilter) {
     return new AutoValue_MeterProviderSharedState(
-        clock, resource, viewRegistry, clock.now(), sampler);
+        clock, resource, viewRegistry, clock.now(), exemplarFilter);
   }
 
   /** Returns the clock used for measurements. */
@@ -42,6 +42,6 @@ public abstract class MeterProviderSharedState {
    */
   abstract long getStartEpochNanos();
 
-  /** Returns the {@link ExemplarSampler} for remembering synchronous measurements. */
-  abstract ExemplarSampler getExemplarSampler();
+  /** Returns the {@link ExemplarFilter} for remembering synchronous measurements. */
+  abstract ExemplarFilter getExemplarFilter();
 }
