@@ -11,7 +11,7 @@ import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.sdk.metrics.common.InstrumentDescriptor;
 import io.opentelemetry.sdk.metrics.common.InstrumentType;
 import io.opentelemetry.sdk.metrics.common.InstrumentValueType;
-import io.opentelemetry.sdk.metrics.exemplar.ExemplarSampler;
+import io.opentelemetry.sdk.metrics.exemplar.ExemplarFilter;
 import io.opentelemetry.sdk.metrics.internal.view.AttributesProcessor;
 import io.opentelemetry.sdk.metrics.internal.view.ViewRegistry;
 import io.opentelemetry.sdk.metrics.view.Aggregation;
@@ -50,7 +50,7 @@ public class AsynchronousMetricStorageTest {
 
     meterProviderSharedState =
         MeterProviderSharedState.create(
-            testClock, Resource.empty(), viewRegistry, ExemplarSampler.builder().build());
+            testClock, Resource.empty(), viewRegistry, ExemplarFilter.sampleWithTraces());
   }
 
   @Test
